+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/NodesResourcesSpec.scala
@@ -8,38 +8,33 @@ import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, OptionValues}
 import pl.touk.nussknacker.engine.additionalInfo.{AdditionalInfo, MarkdownAdditionalInfo}
-import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{
-  BlankId,
-  ExpressionParserCompilationError,
-  InvalidPropertyFixedValue,
-  NodeIdValidationError,
-  ScenarioIdError,
-  ScenarioNameValidationError
-}
+import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{BlankId, ExpressionParserCompilationError, InvalidPropertyFixedValue, NodeIdValidationError, ScenarioIdError, ScenarioNameValidationError}
 import pl.touk.nussknacker.engine.api.displayedgraph.ProcessProperties
 import pl.touk.nussknacker.engine.api.typed.typing
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypingResult}
 import pl.touk.nussknacker.engine.api.{MetaData, ProcessAdditionalFields, StreamMetaData}
 import pl.touk.nussknacker.engine.graph.node.NodeData._
 import pl.touk.nussknacker.engine.graph.evaluatedparam.{BranchParameters, Parameter}
+import pl.touk.nussknacker.engine.api.{ProcessAdditionalFields, StreamMetaData}
+import pl.touk.nussknacker.engine.graph.NodeDataCodec._
+import pl.touk.nussknacker.engine.graph.evaluatedparam.Parameter
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.engine.graph.expression.NodeExpressionId._
 import pl.touk.nussknacker.engine.graph.node
 import pl.touk.nussknacker.engine.graph.node.{Enricher, NodeData}
 import pl.touk.nussknacker.engine.graph.service.ServiceRef
 import pl.touk.nussknacker.engine.graph.sink.SinkRef
-import pl.touk.nussknacker.engine.spel.Implicits._
 import pl.touk.nussknacker.restmodel.definition.UIParameter
 import pl.touk.nussknacker.restmodel.validation.PrettyValidationErrors
 import pl.touk.nussknacker.test.PatientScalaFutures
 import pl.touk.nussknacker.ui.api.helpers.TestFactory.withPermissions
 import pl.touk.nussknacker.ui.api.helpers.{NuResourcesTest, ProcessTestData, TestCategories}
 import pl.touk.nussknacker.ui.process.fragment.FragmentResolver
-import pl.touk.nussknacker.ui.validation.ProcessValidation
+import pl.touk.nussknacker.ui.validation.{NodeValidator, ParametersValidator, UIProcessValidator}
 import pl.touk.nussknacker.engine.kafka.KafkaFactory._
 import pl.touk.nussknacker.ui.suggester.ExpressionSuggester
 
-class NodeResourcesSpec
+class NodesResourcesSpec
     extends AnyFunSuite
     with ScalatestRouteTest
     with FailFastCirceSupport
@@ -54,7 +49,7 @@ class NodeResourcesSpec
 
   private val testProcess = ProcessTestData.sampleDisplayableProcess.copy(category = TestCategories.TestCat)
 
-  private val validation = ProcessValidation(
+  private val validation = UIProcessValidator(
     typeToConfig.mapValues(_.modelData),
     typeToConfig.mapValues(_.scenarioPropertiesConfig),
     typeToConfig.mapValues(_.additionalValidators),
@@ -63,17 +58,19 @@ class NodeResourcesSpec
 
   private val nodeRoute = new NodesResources(
     processService,
-    fragmentRepository,
     typeToConfig.mapValues(_.modelData),
     validation,
+    typeToConfig.mapValues(v => new NodeValidator(v.modelData, fragmentRepository)),
     typeToConfig.mapValues(v =>
       new ExpressionSuggester(
         v.modelData.modelDefinition.expressionConfig,
         v.modelData.modelDefinitionWithTypes.typeDefinitions,
         v.modelData.uiDictServices,
-        v.modelData.modelClassLoader.classLoader
+        v.modelData.modelClassLoader.classLoader,
+        v.scenarioPropertiesConfig.keys
       )
-    )
+    ),
+    typeToConfig.mapValues(v => new ParametersValidator(v.modelData, v.scenarioPropertiesConfig.keys))
   )
 
   private implicit val typingResultDecoder: Decoder[TypingResult] =
@@ -207,45 +204,6 @@ class NodeResourcesSpec
     }
   }
 
-  test("handles global variables in NodeValidationRequest") {
-    saveProcess(testProcess) {
-      val data = node.Join(
-        "id",
-        Some("output"),
-        "enrichWithAdditionalData",
-        List(
-          Parameter("additional data value", "#longValue")
-        ),
-        List(
-          BranchParameters("b1", List(Parameter("role", "'Events'"))),
-          BranchParameters("b2", List(Parameter("role", "'Additional data'")))
-        ),
-        None
-      )
-      val request = NodeValidationRequest(
-        data,
-        ProcessProperties(StreamMetaData()),
-        Map(),
-        Some(
-          Map(
-            // It's a bit tricky, because FE does not distinguish between global and local vars...
-            "b1" -> Map("existButString" -> Typed[String], "meta" -> Typed[MetaData]),
-            "b2" -> Map("longValue" -> Typed[Long], "meta" -> Typed[MetaData])
-          )
-        ),
-        None
-      )
-
-      Post(s"/nodes/${testProcess.id}/validation", toEntity(request)) ~> withPermissions(
-        nodeRoute,
-        testPermissionRead
-      ) ~> check {
-        val res = responseAs[NodeValidationResult]
-        res.validationErrors shouldBe Nil
-      }
-    }
-  }
-
   test("it should return additional info for process properties") {
     saveProcess(testProcess) {
 
@@ -297,7 +255,7 @@ class NodeResourcesSpec
           expressionType = None,
           validationErrors = List(
             PrettyValidationErrors.formatErrorMessage(
-              InvalidPropertyFixedValue("numberOfThreads", Some("Number of threads"), "'a'", List("1", "2"), "")
+              InvalidPropertyFixedValue("numberOfThreads", Some("Number of threads"), "a", List("1", "2"), "")
             )
           ),
           validationPerformed = true
@@ -320,7 +278,7 @@ class NodeResourcesSpec
 
       val expectedErrors = List(
         PrettyValidationErrors.formatErrorMessage(
-          InvalidPropertyFixedValue("numberOfThreads", Some("Number of threads"), "'a'", List("1", "2"), "")
+          InvalidPropertyFixedValue("numberOfThreads", Some("Number of threads"), "a", List("1", "2"), "")
         ),
         PrettyValidationErrors.formatErrorMessage(
           ScenarioNameValidationError(