+++ engine/flink/management/dev-model/src/main/scala/pl/touk/nussknacker/engine/management/sample/service/ConfiguratorService.scala
@@ -3,7 +3,7 @@ package pl.touk.nussknacker.engine.management.sample.service
 import javax.annotation.Nullable
 import javax.validation.constraints.NotBlank
 import pl.touk.nussknacker.engine.api.editor.{DualEditor, DualEditorMode, SimpleEditor, SimpleEditorType}
-import pl.touk.nussknacker.engine.api.validation.Literal
+import pl.touk.nussknacker.engine.api.validation.CompileTimeEvaluableValue
 import pl.touk.nussknacker.engine.api.{MethodToInvoke, ParamName, Service}
 
 import scala.concurrent.Future
@@ -20,7 +20,7 @@ object ConfiguratorService extends Service with Serializable {
       @NotBlank
       template: String,
       @ParamName("Version")
-      @Literal
+      @CompileTimeEvaluableValue
       version: Int,
       @ParamName("JsonConfig")
       @Nullable