+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/parameter/validator/CompileTimeEvaluableValueValidatorExtractor.scala
@@ -1,0 +1,13 @@
+package pl.touk.nussknacker.engine.definition.parameter.validator
+
+import pl.touk.nussknacker.engine.api.definition.{CompileTimeEvaluableValueValidator, ParameterValidator}
+import pl.touk.nussknacker.engine.api.validation.CompileTimeEvaluableValue
+
+object CompileTimeEvaluableValueValidatorExtractor extends ValidatorExtractor {
+
+  override def extract(params: ValidatorExtractorParameters): Option[ParameterValidator] =
+    params.parameterData
+      .getAnnotation[CompileTimeEvaluableValue]
+      .map(_ => CompileTimeEvaluableValueValidator)
+
+}